fix some 'printf' warning.

+++ Common/Source/Devices/devBlueFlyVario.cpp
@@ -248,7 +248,7 @@ namespace dlgBlueFlyConfig {
             }
 
             TCHAR szTmp[50] = {0};
-            _stprintf(szTmp, _T("BlueFlyVario %d/%d"), CurrentPage+1, lstPageWnd.size());
+            _stprintf(szTmp, _T("BlueFlyVario %u/%u"), CurrentPage+1, (unsigned)lstPageWnd.size());
             wfDlg->SetCaption(szTmp);
         }
     }
